+++ package.json
@@ -19,19 +19,17 @@
     "imgup": "./dist/index.js"
   },
   "dependencies": {
+    "@types/axios": "^0.14.0",
     "@types/form-data": "^2.5.0",
     "@types/lodash": "^4.14.176",
     "@types/node": "^16.11.7",
-    "axios": "^0.24.0",
+    "axios": "^1.2.6",
     "chalk": "^4.1.2",
     "commander": "^8.3.0",
-    "dotenv": "^10.0.0",
     "form-data": "^4.0.0",
+    "husky": "^4.3.8",
     "lodash": "^4.17.21",
     "ora": "^6.0.1",
     "prettier": "^2.4.1"
-  },
-  "devDependencies": {
-    "husky": "^4.3.8"
   }
 }
